Upload product from Vendor app

+++ mac_store_app/lib/models/cart_models.dart
@@ -1,6 +1,6 @@
 class CartModels {
   final String productName;
-  final int productPrice;
+  final double productPrice;
   final String categoryName;
   final List imageUrl;
   int quantity; //total amount buyer wants to but
